beginning aggregation buffer
