+++ lake/top/lake_top.py
@@ -1,6 +1,7 @@
 from kratos import *
 from lake.modules.passthru import *
 from lake.modules.sram_stub import SRAMStub
+from lake.modules.aggregation_buffer import AggregationBuffer
 
 class LakeTop(Generator):
     def __init__(
@@ -14,16 +15,38 @@ class LakeTop(Generator):
                 interconnect_output_ports = 1,
                 mem_input_ports = 1,
                 mem_output_ports = 1,
-                use_sram_stub = 1
+                use_sram_stub = 1,
+                agg_height = 1,
+                transpose_height = 1,
+                max_agg_schedule = 12
                 ):
         super().__init__("LakeTop")
-        self._in = self.input("i_data_in", width)
-        self._out = self.output("o_data_out", width)
+
+        self.data_width = data_width
+        self.mem_width = mem_width
+        self.mem_depth = mem_depth
+        self.banks = banks
+        self.iterator_support = iterator_support
+        self.interconnect_input_ports = interconnect_input_ports
+        self.interconnect_output_ports = interconnect_output_ports
+        self.mem_input_ports = mem_input_ports
+        self.mem_output_ports = mem_output_ports
+        self.use_sram_stub = use_sram_stub
+        self.agg_height = agg_height
+        self.transpose_height = transpose_height
+
         self._clk = self.clock("i_clk")
         self._rst_n = self.reset("i_rst_n")
 
+        # Get the input ports from the interconnect
+        self._data_in = self.input("i_data_in", self.data_width, size=self.interconnect_input_ports, packed=True, explicit_array=True)
+        self._valid_in = self.input("i_valid_in", self.interconnect_input_ports, packed=True, explicit_array=True)
+
+        self._data_out = self.output("o_data_out", self.data_width, size=self.interconnect_output_ports, packed=True, explicit_array=True)
+        self._valid_out = self.output("o_valid_out", self.interconnect_output_ports, packed=True, explicit_array=True)
+
         # First wrap sram_stub
-        sram_stub = SRAMStub(width, 1024)
+        sram_stub = SRAMStub(mem_width, 1024)
         self.add_child_generator(f"u_sram_stub_0", sram_stub)
         self.wire(sram_stub.i_data, self._in)
         self.wire(self._out, sram_stub.o_data)
@@ -34,9 +57,12 @@ class LakeTop(Generator):
         self.wire(sram_stub.i_clk, self._clk)
         self.wire(sram_stub.i_rst_n, self._rst_n)
 
-
         # Add input aggregations buffers
-
+        for i in range(self.interconnect_input_ports):
+            # add children aggregator buffers...
+            self.add_child(f"agg_in_{i}", AggregationBuffer(self.agg_height, self.data_width, self.mem_width))
+            # Also add aggregation buffer config nodes...?
+            # now wire it up
 
 
         # Add transpose buffers at output
